Improve logging, clean code

+++ app/utils.py
@@ -1,6 +1,7 @@
 import os
 import sys
 import json
+import logging
 import tempfile
 import xmltodict
 import subprocess
@@ -10,6 +11,9 @@ from random import randint
 from seiscomp.seismology import Regions
 from urllib.request import Request, urlopen
 
+DEBUG = False
+
+logger = logging.getLogger(__name__)
 
 def load_config():
     curr_dir = os.path.dirname(os.path.abspath(__file__))
@@ -22,7 +26,6 @@ def load_config():
     os.chmod(commit_script, 0o755)
     return config
 
-DEBUG = False
 CONFIG = load_config()
 
 def update_config(content):
@@ -168,13 +171,14 @@ def commit_script(qml):
     curr_dir = os.path.dirname(os.path.abspath(__file__))
     commit_script = os.path.join(curr_dir, '..', 'commit_script.sh')
     _, qml_filename = tempfile.mkstemp(suffix='.xml')
-    sys.stderr.write(f'{qml_filename}\n')
+    logger.debug(f'commit quakeml file: {qml_filename}\n')
     with open(qml_filename, 'wb') as f:
         f.write(qml)
     p = subprocess.Popen([commit_script, qml_filename],
                          stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     _, error_message = p.communicate()
-    os.remove(qml_filename)
+    if not DEBUG:
+        os.remove(qml_filename)
     return {
         'message': error_message.decode('utf-8'),
         'return_code': p.returncode
